requirements.txt
# Core Frameworks
fastapi>=0.100.0 
pydantic>=2.0
uvicorn
streamlit
streamlit-audiorecorder

# AI & Machine Learning
openai
google-generativeai
soundfile
# google-cloud-texttospeech
transformers
torch
scipy # For saving wav files


# Utilities
python-dotenv
requests

python-multipart